refactor: moving evaluation process to module folder
