+++ src/app/modules/reports/views/evaluation-process/evaluation-process-form/evaluation-process-form.component.scss
@@ -1,4 +1,4 @@
-@use '../../../../styles/utils/index.scss' as main-styles;
+@use '../../../../../../styles/utils/index.scss' as main-styles;
 
 .modal {
   padding-top: 25px;
