+++ src/app/modules/reports/views/evaluation-process/evaluation-process-list/evaluation-process-list.component.scss
@@ -1,4 +1,4 @@
-@use './../../../../styles/utils/index.scss' as main-styles;
+@use '../../../../../../styles/utils/index.scss' as main-styles;
 
 .noEvalsContainer {
   display: flex;
