+++ src/app/modules/reports/views/evaluation-process/evaluation-process-list/evaluation-process-list.component.spec.ts
@@ -5,7 +5,7 @@ import { provideHttpClient } from '@angular/common/http';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { provideNoopAnimations } from '@angular/platform-browser/animations';
 import { of } from 'rxjs';
-import { EvaluationsService } from '../../../core/services/api/evaluations.service';
+import { EvaluationsService } from '../../../../../core/services/api/evaluations.service';
 
 describe('EvaluationProcessListComponent', () => {
   let component: EvaluationProcessListComponent;
