+++ src/app/modules/reports/views/evaluation-process/evaluation-process-list/evaluation-process-list.component.ts
@@ -10,14 +10,14 @@ import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { MatTable, MatTableModule } from '@angular/material/table';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { Router } from '@angular/router';
-import { GENERAL_MESSAGES } from '../../../core/constants/messages';
-import { PagedReadEvaluationProcess } from '../../../core/models/evaluation-request.model';
-import { EvaluationModel } from '../../../core/models/evaluation.model';
-import { ModalComponent } from '../../../shared/components/modal-dialog/modal-dialog.component';
+import { GENERAL_MESSAGES } from '../../../../../core/constants/messages';
+import { PagedReadEvaluationProcess } from '../../../../../core/models/evaluation-request.model';
+import { EvaluationModel } from '../../../../../core/models/evaluation.model';
+import { ModalComponent } from '../../../../../shared/components/modal-dialog/modal-dialog.component';
 import { EvaluationProcessFormComponent } from '../evaluation-process-form/evaluation-process-form.component';
-import { EmptyDataComponent } from '../../../shared/components/empty-data/empty-data.component';
-import { EvaluationsService } from '../../../core/services/api/evaluations.service';
-import { Status } from '../../../core/constants/common';
+import { EmptyDataComponent } from '../../../../../shared/components/empty-data/empty-data.component';
+import { EvaluationsService } from '../../../../../core/services/api/evaluations.service';
+import { Status } from '../../../../../core/constants/common';
 
 @Component({
   selector: 'app-evaluation-process-list',
